+++ packages/input-subnet/index.js
@@ -1,4 +1,4 @@
-import VESubnet from './src/SubnetMask.vue';
+import VESubnet from './src/main.vue';
 
 VESubnet.install = function (Vue) {
   Vue.component(VESubnet.name, VESubnet);