+++ packages/tool-tip/src/main.vue
@@ -1,9 +1,9 @@
 <template>
   <transition :name="transition">
-    <div :class="['v-easy-tip', 've-tip-'+placement, Class]" v-show="domVisible" @mouseenter="enter" @mouseleave="leave">
-      <p v-if="html" v-html="html"></p>
+    <div :class="['v-easy-tip', 've-tip-'+placement, 've-tip--' + effect, Class]" v-show="domVisible" @mouseenter="enter" @mouseleave="handleMouseLeave" @click="handleClick">
+      <p v-if="html" v-html="html" />
       <p v-else-if="content">{{ content }}</p>
-      <render-node v-else></render-node>
+      <render-node v-else />
     </div>
   </transition>
 </template>
@@ -21,17 +21,36 @@
         hideAfter: 200,
         transition: 'fade',
         enterable: true,
-        html: ''
+        target: '',
+        html: '',
+        effect: 'dark',
       };
     },
     components: {
       renderNode: {
         render (h) {
-          return this.$parent.VNode()
-        }
+          const parent = this.$parent;
+          let VNode = h('span');
+          if (typeof parent.VNode === 'function') {
+            VNode = parent.VNode()
+          } else {
+            console.warn('VNode is not a function')
+          }
+          return VNode
+        },
       }
     },
     methods: {
+      handleClick(event) {
+        event.stopPropagation();
+      },
+      handleMouseLeave() {
+        if (this.target === 'click') {
+
+        } else {
+          this.leave()
+        }
+      },
       enter() {
         if (!this.enterable) return false;
         this.hover = true;