+++ src/utils/icon-style.js
@@ -4,6 +4,7 @@ export const computedIconStyle = function computedIconStyle (style) {
     case 'brands': className = 'fab'; break;
     case 'regular': className = 'far'; break;
     case 'solid': className = 'fa'; break;
+    default: className = style;
   }
   return className;
 };